+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.1.2 [unreleased]
 
 - Instance Profiles: List them; attach them to an instance. Ex: `attach instanceprofiles name=...`, `awless ls instanceprofiles`
+- Replace in one command an InstanceProfile on a given instance with the `replace=true` param. Ex: `attach instanceprofile .... replace=true`
 
 ### Features